% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate.R
\name{orderly2outpack}
\alias{orderly2outpack}
\title{Migrate orderly to outpack}
\usage{
orderly2outpack(src, dest)
}
\arguments{
\item{src}{Path to the orderly archive to migrate. This must be
"complete" (i.e., containing a full graph and no metadata only
pulls)}

\item{dest}{The destination to create a new outpack archive. This
will be created with no user-visible archive directory, with a
file store, and requiring a full tree (i.e., in a suitable
configuration to use as an upstream source).}
}
\value{
The path to the newly created archive
}
\description{
Migrate an orderly archive to outpack. This function is subject to
change, and does not handle all failure modes. It is quite slow
because it involves hashing the contents of your archive several
times - if you have GBs of files expect this to take minutes or
hours.
}
