% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate.R
\name{orderly2outpack}
\alias{orderly2outpack}
\title{Migrate orderly to outpack}
\usage{
orderly2outpack(src, dest, link = FALSE, keep_going = FALSE, parallel = FALSE)
}
\arguments{
\item{src}{Path to the orderly archive to migrate. This must be
"complete" (i.e., containing a full graph and no metadata only
pulls)}

\item{dest}{The destination to create a new outpack archive. This
will be created with no user-visible archive directory, with a
file store, and requiring a full tree (i.e., in a suitable
configuration to use as an upstream source).}

\item{link}{Logical, indicating if the file store should be hard
links to the existing orderly archive, rather than copies of the
files. Doing this carries some risk and constraints: must be on
the same file system, can't make files readonly, changes
propagated both ways. As a result, only do this where you are
confident that nothing will alter files in the source archive!
The use case here is in continually migrating an orderly
repository.}

\item{keep_going}{Logical, indicating if we should try and migrate as
much as possible, skipping over packets that fail.}

\item{parallel}{Logical, indicating if we should process data in
parallel.  This uses \code{parallel::mclapply} so it only works on
Linux.  Use the \code{mc.cores} option or \code{MC_CORES} environment
variable to control the number of cores used.}
}
\value{
The path to the newly created archive
}
\description{
Migrate an orderly archive to outpack. This function is subject to
change, and does not handle all failure modes. It is quite slow
because it involves hashing the contents of your archive several
times - if you have GBs of files expect this to take minutes or
hours.
}
