% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{orderly2outpack_src}
\alias{orderly2outpack_src}
\title{Migrate orderly source}
\usage{
orderly2outpack_src(path, delete_yml = FALSE, strict = FALSE)
}
\arguments{
\item{path}{Path to the orderly source directory}

\item{delete_yml}{Logical, indicating if the \code{orderly.yml} files
should be deleted.}

\item{strict}{Logical, indicating if we should enable
\code{\link[orderly3:orderly_strict_mode]{orderly3::orderly_strict_mode()}} in the resulting source
files}
}
\value{
Nothing, called for side effects only
}
\description{
Migrate an orderly \emph{source directory} for use with outpack.  This
works in place, writing out \code{orderly.R} files and deleting
\code{orderly.yml} files (optionally). This is intended to be run on a
clean clone of your source git repository.
}
